requirements.txt
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.2.2
xgboost==1.7.6
matplotlib==3.8.0
seaborn==0.12.2
plotly==5.15.0
streamlit==1.26.0
joblib==1.3.2
jupyterlab==4.3.0
openpyxl==3.1.2
python-pptx==0.6.21
pillow==10.0.0
imbalanced-learn

